# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nordigen Account Information Services API
  version: 2.0 (v2)
servers:
- url: /
paths:
...
  /api/v2/requisitions/:
    get:
      description: API endpoints related to requisitions.
      operationId: retrieve all requisitions
      parameters:
      - description: Number of results to return per page.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
      - description: The initial index from which to return the results.
        explode: true
        in: query
        name: offset
        required: false
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRequisitionV2List'
          description: retrieve all requisitions
      security:
      - tokenAuth: []
      - cookieAuth: []
      tags:
      - requisitions
    post:
      description: API endpoints related to requisitions.
      operationId: requisition created
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequisitionV2'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RequisitionV2'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RequisitionV2'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpectacularRequisitionV2'
          description: Requisition has been succesfully created
        "400":
          content:
            application/json:
              examples:
                ReferenceFieldErrors:
                  description: Errors related to 'reference' field.
                  summary: Reference field errors
                  value:
                    reference:
                      summary: Client Reference must be unique
                      detail: 'Client Reference: $REFERENCE_ID already exists'
                    status_code: 400
                UserLanguageFieldErrors:
                  description: Errors related to 'user_language' field.
                  summary: User language field errors
                  value:
                    user_language:
                      summary: Provided user_language is invalid or not supported
                      detail: '''$LNG_CODE'' is invalid or not supported language'
                    status_code: 400
                AgreementFieldErrors:
                  description: Errors related to 'agreement' field.
                  summary: Agreement field errors
                  value:
                    agreement:
                    - summary: Unknown Institution ID $ASPSP_ID
                      detail: Get Institution IDs from /institutions/?country={$COUNTRY_CODE}
                    - summary: Invalid EndUserAgreement ID
                      detail: $AGREEMENT_ID is not a valid UUID. Please specify valid
                        agreement from /api/agreements/enduser/?={$ENDUSER_ID} or
                        create a new one
                    status_code: 400
                RedirectFieldErrors:
                  description: Errors related to 'redirect' field.
                  summary: Redirect field errors
                  value:
                    redirect:
                    - summary: Redirect URL is required
                      detail: Where an end user will be redirected after finishing
                        authentication in ASPSP
                    - summary: Invalid redirect URI
                      detail: Redirect URI must have a valid URI structure
                    status_code: 400
                SSNFieldErrors:
                  description: Errors related to 'ssn' field.
                  summary: SSN field errors
                  value:
                    ssn:
                      summary: SSN verification is not supported
                      detail: SSN verification is not supported for $INSTITUTION_ID
                    status_code: 400
                AccountSelectionFieldErrors:
                  description: Errors related to 'account_selection' field.
                  summary: Account selection field errors
                  value:
                    account_selection:
                      summary: Account selection is not supported
                      detail: Account selection is not supported for $INSTITUTION_ID
                    status_code: 400
                InstitutionFieldErrors:
                  description: Errors related to 'institution' field.
                  summary: Institution field errors
                  value:
                    institution_id:
                      summary: Unknown Institution ID $INSTITUTION_ID
                      detail: Get Institution IDs from /institutions/?country={$COUNTRY_CODE}
                    status_code: 400
                UnknownFields:
                  description: Unknown Fields
                  summary: Unknown Fields
                  value:
                    summary: Unknown fields
                    detail: Unknown fields {${FIELD}} in payload
                    status_code: 400
                FieldsRequired:
                  description: Field is required
                  summary: Fields required
                  value:
                    redirect:
                    - This field is required.
                    institution_id:
                    - This field is required.
                    status_code: 400
              schema:
                additionalProperties: {}
                type: object
        "404":
          content:
            application/json:
              examples:
                AgreementNotFoundErrors:
                  description: Errors related to 'agreement' field.
                  summary: Agreement not found errors
                  value:
                    agreement:
                      summary: 'Agreement: $AGREEMENT_ID not found'
                      detail: Please specify valid agreement from /agreements/enduser/?={enduser_id}
                        or create a new one
                    status_code: 404
              schema:
                additionalProperties: {}
                type: object
      security:
      - tokenAuth: []
      - cookieAuth: []
      tags:
      - requisitions
  /api/v2/requisitions/{id}/:
    delete:
      description: Delete Requisition and all End User Agreements.
      operationId: delete requisition by id v2
      parameters:
      - description: A UUID string identifying this requisition.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          description: requisition has been deleted
        "400":
          content:
            application/json:
              examples:
                InvalidID:
                  description: Invalid ID
                  summary: Invalid ID
                  value:
                    summary: Invalid ID
                    detail: foo-bar-baz is not a valid UUID.
                    status_code: 400
              schema:
                additionalProperties: {}
                type: object
        "404":
          content:
            application/json:
              examples:
                NotFoundError:
                  description: Not found error
                  summary: Not found error
                  value:
                    detail: Not found.
                    status_code: 404
              schema:
                additionalProperties: {}
                type: object
      security:
      - tokenAuth: []
      - cookieAuth: []
      tags:
      - requisitions
    get:
      description: API endpoints related to requisitions.
      operationId: requisition by id
      parameters:
      - description: A UUID string identifying this requisition.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequisitionV2'
          description: Get requisition by id
        "404":
          content:
            application/json:
              examples:
                NotFoundError:
                  description: Not found error
                  summary: Not found error
                  value:
                    detail: Not found.
                    status_code: 404
              schema:
                additionalProperties: {}
                type: object
      security:
      - tokenAuth: []
      - cookieAuth: []
      tags:
      - requisitions
...
components:
  schemas:
...
    PaginatedRequisitionV2List:
      example:
        next: https://ob.nordigen.com/api/v2/requisitions/?limit=100&offset=0
        previous: https://ob.nordigen.com/api/v2/requisitions/?limit=100&offset=0
        count: 123
        results:
        - redirect: https://openapi-generator.tech
          reference: reference
          account_selection: false
          agreement: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          created: 2000-01-23T04:56:07.000+00:00
          link: https://openapi-generator.tech
          user_language: user_language
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          accounts:
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: ""
          institution_id: institution_id
          ssn: ssn
        - redirect: https://openapi-generator.tech
          reference: reference
          account_selection: false
          agreement: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          created: 2000-01-23T04:56:07.000+00:00
          link: https://openapi-generator.tech
          user_language: user_language
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          accounts:
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: ""
          institution_id: institution_id
          ssn: ssn
      properties:
        count:
          example: 123
          type: integer
        next:
          example: https://ob.nordigen.com/api/v2/requisitions/?limit=100&offset=0
          format: uri
          nullable: true
          type: string
        previous:
          example: https://ob.nordigen.com/api/v2/requisitions/?limit=100&offset=0
          format: uri
          nullable: true
          type: string
        results:
          items:
            $ref: '#/components/schemas/RequisitionV2'
          type: array
      type: object
...
    RequisitionV2:
      description: Get requisition by id.
      example:
        redirect: https://openapi-generator.tech
        reference: reference
        account_selection: false
        agreement: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        created: 2000-01-23T04:56:07.000+00:00
        link: https://openapi-generator.tech
        user_language: user_language
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        accounts:
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: ""
        institution_id: institution_id
        ssn: ssn
      properties:
        id:
          format: uuid
          readOnly: true
          type: string
        created:
          description: The date & time at which the requisition was created.
          format: date-time
          nullable: true
          readOnly: true
          title: Created Date
          type: string
        redirect:
          description: redirect URL to your application after end-user authorization
            with ASPSP
          format: uri
          nullable: true
          type: string
        status:
          allOf:
          - $ref: '#/components/schemas/Status6e6Enum'
          description: status of this requisition
          readOnly: true
          title: Requisition status
        institution_id:
          description: an Institution ID for this Requisition
          type: string
        agreement:
          description: EUA associated with this requisition
          format: uuid
          type: string
        reference:
          description: additional ID to identify the end user
          maxLength: 256
          type: string
        accounts:
          description: array of account IDs retrieved within a scope of this requisition
          items:
            format: uuid
            title: accounts
            type: string
          readOnly: true
          type: array
        user_language:
          description: A two-letter country code (ISO 639-1)
          maxLength: 5
          type: string
        link:
          default: https://ob.nordigen.com/psd2/start/3fa85f64-5717-4562-b3fc-2c963f66afa6/{$INSTITUTION_ID}
          description: link to initiate authorization with Institution
          format: uri
          readOnly: true
          type: string
        ssn:
          description: optional SSN field to verify ownership of the account
          maxLength: 64
          type: string
        account_selection:
          default: false
          description: option to enable account selection view for the end user
          type: boolean
      required:
      - institution_id
      - redirect
      type: object
...
    SpectacularRequisitionV2:
      description: Create requisition.
      example:
        redirect: https://openapi-generator.tech
        reference: reference
        account_selection: false
        agreement: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        created: 2000-01-23T04:56:07.000+00:00
        link: https://openapi-generator.tech
        user_language: user_language
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        accounts:
        - ""
        - ""
        status: ""
        institution_id: institution_id
        ssn: ssn
      properties:
        id:
          format: uuid
          readOnly: true
          type: string
        created:
          description: The date & time at which the requisition was created.
          format: date-time
          nullable: true
          readOnly: true
          title: Created Date
          type: string
        redirect:
          description: redirect URL to your application after end-user authorization
            with ASPSP
          format: uri
          nullable: true
          type: string
        status:
          allOf:
          - $ref: '#/components/schemas/Status6e6Enum'
          description: status of this requisition
          readOnly: true
          title: Requisition status
        institution_id:
          description: an Institution ID for this Requisition
          type: string
        agreement:
          description: EUA associated with this requisition
          format: uuid
          type: string
        reference:
          description: additional ID to identify the end user
          maxLength: 256
          type: string
        accounts:
          default: []
          description: array of account IDs retrieved within a scope of this requisition
          items: {}
          readOnly: true
          type: array
        user_language:
          description: A two-letter country code (ISO 639-1)
          maxLength: 5
          type: string
        link:
          default: https://ob.nordigen.com/psd2/start/3fa85f64-5717-4562-b3fc-2c963f66afa6/{$INSTITUTION_ID}
          description: link to initiate authorization with Institution
          format: uri
          readOnly: true
          type: string
        ssn:
          description: optional SSN field to verify ownership of the account
          maxLength: 64
          type: string
        account_selection:
          default: false
          description: option to enable account selection view for the end user
          type: boolean
      required:
      - institution_id
      - redirect
      type: object
    Status6e6Enum:
      enum:
      - '{"short":"CR","long":"CREATED","description":"Requisition has been successfully
        created"}'
      - '{"short":"LN","long":"LINKED","description":"Account has been successfully
        linked to requisition"}'
      - '{"short":"EX","long":"EXPIRED","description":"Access to account has expired
        as set in End User Agreement"}'
      - '{"short":"RJ","long":"REJECTED","description":"SSN verification has failed"}'
      type: string
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: Session
      type: apiKey
    jwtAuth:
      bearerFormat: Bearer
      scheme: bearer
      type: http
    tokenAuth:
      description: Token-based authentication with required prefix "Token"
      in: header
      name: Authorization
      type: apiKey