# go/model_paginated_requisition_v2_list.go
/*
 * Nordigen Account Information Services API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 2.0 (v2)
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type PaginatedRequisitionV2List struct {

	Count int32 `json:"count,omitempty"`

	Next *string `json:"next,omitempty"`

	Previous *string `json:"previous,omitempty"`

	Results []RequisitionV2 `json:"results,omitempty"`
}

// AssertPaginatedRequisitionV2ListRequired checks if the required fields are not zero-ed
func AssertPaginatedRequisitionV2ListRequired(obj PaginatedRequisitionV2List) error {
	for _, el := range obj.Results {
		if err := AssertRequisitionV2Required(el); err != nil {
			return err
		}
	}
	return nil
}

// AssertRecursePaginatedRequisitionV2ListRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of PaginatedRequisitionV2List (e.g. [][]PaginatedRequisitionV2List), otherwise ErrTypeAssertionError is thrown.
func AssertRecursePaginatedRequisitionV2ListRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aPaginatedRequisitionV2List, ok := obj.(PaginatedRequisitionV2List)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertPaginatedRequisitionV2ListRequired(aPaginatedRequisitionV2List)
	})
}